pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapi-to-mcp"
version = "0.3.0"
description = "Turn any OpenAPI/Swagger spec into a runnable MCP server in one command."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "openapi-to-mcp contributors" }
]
license = "MIT"
dependencies = [
  "httpx>=0.27.0",
  "mcp>=1.20.0",
  "PyYAML>=6.0.1",
  "typer>=0.12.0",
  "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0"
]

[project.scripts]
openapi-to-mcp = "openapi_to_mcp.cli:app"

[project.urls]
Homepage = "https://github.com/vect-G/openapi-to-mcp"
Repository = "https://github.com/vect-G/openapi-to-mcp"
Issues = "https://github.com/vect-G/openapi-to-mcp/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
